pyproject.toml
[project]
name = "aurora_backend_llm"
version = "0.1.0"
description = "aurora_backend_llm using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.114.0,<1.0.0",
    "flask>=2.3.0",
    "flask-sqlalchemy>=3.1.0",
    "flask-migrate>=4.0.0",
    "flask-jwt-extended>=4.5.0",
    "psycopg2-binary>=2.9.0",
    "passlib>=1.7.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "email-validator>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0"
]

[project.scripts]
aurora_backend_llm = "aurora_backend_llm.main:run"
run_crew = "aurora_backend_llm.main:run"
train = "aurora_backend_llm.main:train"
replay = "aurora_backend_llm.main:replay"
test = "aurora_backend_llm.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
